include hlsl, Gui outline, file Save&Load 추가 Tuto, player, playerBullet, playerSpecialAttack 수정

+++ D2D/Bullet/PlayerBulletManager.h
@@ -15,7 +15,7 @@ public:
 
 	int GetNextIndex() { return nextIndex; }
 	float GetTime() { return time; }
-	float GetLastIndex() { return lastIndex; }
+	int GetLastIndex() { return lastIndex; }
 	float GetSpeakerSpeed() { return speakerSpeed; }
 	shared_ptr<PlayerBullet> GetBullet(int index) { return bullets[index]; }
